feat(driver-adapters): extracted platform-agnostic "DriverAdapterError" into "driver_adapters::error"
