+++ query-engine/driver-adapters/src/napi/result.rs
@@ -1,49 +1,7 @@
 use napi::{bindgen_prelude::FromNapiValue, Env, JsUnknown, NapiValue};
-use quaint::error::{Error as QuaintError, ErrorKind, MysqlError, PostgresError, SqliteError};
-use serde::Deserialize;
+use quaint::error::{Error as QuaintError, ErrorKind};
 
-#[derive(Deserialize)]
-#[serde(remote = "PostgresError")]
-pub struct PostgresErrorDef {
-    code: String,
-    message: String,
-    severity: String,
-    detail: Option<String>,
-    column: Option<String>,
-    hint: Option<String>,
-}
-
-#[derive(Deserialize)]
-#[serde(remote = "MysqlError")]
-pub struct MysqlErrorDef {
-    pub code: u16,
-    pub message: String,
-    pub state: String,
-}
-
-#[derive(Deserialize)]
-#[serde(remote = "SqliteError", rename_all = "camelCase")]
-pub struct SqliteErrorDef {
-    pub extended_code: i32,
-    pub message: Option<String>,
-}
-
-#[derive(Deserialize)]
-#[serde(tag = "kind")]
-/// Wrapper for JS-side errors
-pub(crate) enum DriverAdapterError {
-    /// Unexpected JS exception
-    GenericJs {
-        id: i32,
-    },
-    UnsupportedNativeDataType {
-        #[serde(rename = "type")]
-        native_type: String,
-    },
-    Postgres(#[serde(with = "PostgresErrorDef")] PostgresError),
-    Mysql(#[serde(with = "MysqlErrorDef")] MysqlError),
-    Sqlite(#[serde(with = "SqliteErrorDef")] SqliteError),
-}
+use crate::error::DriverAdapterError;
 
 impl FromNapiValue for DriverAdapterError {
     unsafe fn from_napi_value(napi_env: napi::sys::napi_env, napi_val: napi::sys::napi_value) -> napi::Result<Self> {
